2.0: minor updates

+++ src/xpn_client/nfi/nfi_tcp_server/tcp_server_comm.c
@@ -362,7 +362,7 @@ ssize_t tcp_server_write_data(int fd, char *data, ssize_t size, char *id){
 
 
 #ifdef DBG_COMM
-	printf("[NFI_COMM]client: write_data(%d): %d = %d ID=%s --th:%d--\n",fd,size,ret,id,(int)pthread_self());
+	printf("[NFI_COMM]client: write_data(%d): %lu = %d ID=%s --th:%d--\n", fd, (unsigned long)size, ret, id, (int)pthread_self());
 #endif
                 if(ret <= 0){                                                                                               
                         perror("client: Error write_comm:");                                                                 
@@ -372,13 +372,13 @@ ssize_t tcp_server_write_data(int fd, char *data, ssize_t size, char *id){
 	
 	if(ret == -1){
 #ifdef DBG_COMM
-		printf("[NFI_COMM]client: write_data(%d): err %d  ID=%s --th:%d--\n",fd,ret,id,(int)pthread_self());
+		printf("[NFI_COMM]client: write_data(%d): err %d  ID=%s --th:%d--\n", fd, ret, id, (int)pthread_self());
 #endif	
 		perror("client: write_data");
 		return ret;
 	}
 #ifdef DBG_COMM
-	printf("[NFI_COMM]client: write_data(%d): %d de %d ID=%s --th:%d--\n",fd,cont,size,id,(int)pthread_self());
+	printf("[NFI_COMM]client: write_data(%d): %d de %lu ID=%s --th:%d--\n", fd, cont, (unsigned long)size, id, (int)pthread_self());
 #endif	
 #ifdef DBG_COMM
 	tcp_server_write_data_test(fd, id);
